Revisão

+++ aula05/ex06.cpp
@@ -1,13 +1,13 @@
 #include <iostream>
 using namespace std;
 
-void metaDobro(int valor, int *met, int *dob){
+void metaDobro(float valor, float *met, float *dob){
     *met = valor / 2;
     *dob = valor * 2;
 }
 
 int main(){
-    int a, b, c;
+    float a, b, c;
     cout << "Digite um valor ";
     cin >> a;
     metaDobro(a, &b, &c);
